+++ integration_tests/juniper_tests/src/codegen/derive_object.rs
@@ -79,7 +79,7 @@ struct WithCustomContext {
     a: bool,
 }
 
-#[juniper::object]
+#[juniper::graphql_object]
 impl Query {
     fn obj() -> Obj {
         Obj {