+++ juniper/src/schema/schema.rs
@@ -106,7 +106,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__Schema"
     Context = SchemaType<'a, S>,
     Scalar = S,
@@ -145,7 +145,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__Type"
     Context = SchemaType<'a, S>,
     Scalar = S,
@@ -278,7 +278,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__Field",
     Context = SchemaType<'a, S>,
     Scalar = S,
@@ -317,7 +317,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__InputValue",
     Context = SchemaType<'a, S>,
     Scalar = S,
@@ -346,7 +346,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__EnumValue",
     Scalar = S,
     // FIXME: make this redundant.
@@ -373,7 +373,7 @@ where
     }
 }
 
-#[crate::object_internal(
+#[crate::graphql_object_internal(
     name = "__Directive",
     Context = SchemaType<'a, S>,
     Scalar = S,