+++ juniper_codegen/src/lib.rs
@@ -135,7 +135,7 @@ More advanced use cases are introduced step by step.
 struct Query;
 
 // We prefix the impl Block with the procedural macro.
-#[juniper::object]
+#[juniper::graphql_object]
 impl Query {
 
     // A **warning**: only GraphQL fields can be specified in this impl block.
@@ -186,7 +186,7 @@ impl Person {
     }
 }
 
-#[juniper::object]
+#[juniper::graphql_object]
 impl Person {
     fn first_name(&self) -> &str {
         &self.first_name
@@ -226,7 +226,7 @@ impl juniper::Context for Context {}
 
 struct Query;
 
-#[juniper::object(
+#[juniper::graphql_object(
     // Here we specify the context type for this object.
     Context = Context,
 )]
@@ -256,7 +256,7 @@ struct InternalQuery;
 // Doc comments can be used to specify graphql documentation.
 /// GRAPHQL DOCUMENTATION.
 /// More info for GraphQL users....
-#[juniper::object(
+#[juniper::graphql_object(
     // You can rename the type for GraphQL by specifying the name here.
     name = "Query",
     // You can also specify a description here.
@@ -319,7 +319,7 @@ struct WithLifetime<'a> {
     value: &'a str,
 }
 
-#[juniper::object]
+#[juniper::graphql_object]
 impl<'a> WithLifetime<'a> {
     fn value(&self) -> &str {
         self.value
@@ -340,7 +340,7 @@ You can easily specify a custom scalar though.
 
 struct Query;
 
-#[juniper::object(
+#[juniper::graphql_object(
     Scalar = MyCustomScalar,
 )]
 impl Query {
@@ -358,7 +358,7 @@ struct User {
     r#type: String,
 }
 
-#[juniper::object]
+#[juniper::graphql_object]
 impl User {
     fn r#type(&self) -> &str {
         &self.r#type
@@ -368,7 +368,7 @@ impl User {
 
 */
 #[proc_macro_attribute]
-pub fn object(args: TokenStream, input: TokenStream) -> TokenStream {
+pub fn graphql_object(args: TokenStream, input: TokenStream) -> TokenStream {
     let gen = impl_object::build_object(args, input, false);
     gen.into()
 }
@@ -376,7 +376,7 @@ pub fn object(args: TokenStream, input: TokenStream) -> TokenStream {
 /// A proc macro for defining a GraphQL object.
 #[doc(hidden)]
 #[proc_macro_attribute]
-pub fn object_internal(args: TokenStream, input: TokenStream) -> TokenStream {
+pub fn graphql_object_internal(args: TokenStream, input: TokenStream) -> TokenStream {
     let gen = impl_object::build_object(args, input, true);
     gen.into()
 }